Find work_packages and working_hours in CustomerArea
Before, nowhere has been able to see the particular working_packages
and working_hours linked to a project.
Change to wrong working_hour on a invoice was tricky.

Now, you can edit, copy and delete every work_package linked to a
project,
and  edit and delete every working_hour linked to a work_package.

Much more easier to change a working_hour that isn't correct allocated
in a invoice.